+++ pcdsdaq/ext_scripts.py
@@ -27,13 +27,32 @@ def call_script(args, timeout=None, ignore_return_code=False):
         raise
 
 
-def hutch_name():
+cache = {}
+
+
+def cache_script(args, timeout=None, ignore_return_code=False):
+    key = ' '.join(args)
+    try:
+        return cache[key]
+    except KeyError:
+        output = call_script(args, timeout=timeout,
+                             ignore_return_code=ignore_return_code)
+        cache[key] = output
+        return output
+
+
+def clear_script_cache():
+    global cache
+    cache = {}
+
+
+def hutch_name(timeout=10):
     script = SCRIPTS.format('latest', 'get_hutch_name')
-    name = call_script(script)
+    name = cache_script(script, timeout=timeout)
     return name.lower().strip(' \n')
 
 
-def get_run_number(hutch=None, live=False):
+def get_run_number(hutch=None, live=False, timeout=1):
     latest = hutch or 'latest'
     script = SCRIPTS.format(latest, 'get_lastRun')
     args = [script]
@@ -41,25 +60,36 @@ def get_run_number(hutch=None, live=False):
         args += ['-i', hutch]
     if live:
         args += ['-l']
-    run_number = call_script(args)
+    run_number = call_script(args, timeout=timeout)
     return int(run_number)
 
 
-def get_ami_proxy(hutch):
-    # This is mostly copied from old hutch python verbatim
-    # I don't have useful explanations for what these regular expressions
-    # are used for
+def get_ami_proxy(hutch, timeout=10):
+    """
+    Match the output text from procmgr ami status.
+
+    The line we're looking for always includes the text ami_proxy.
+    The -I argument holds the IP address or hostname of the ami proxy.
+
+    I thought the first host in the list was the name of the ami proxy, but
+    this does not seem to be consistent with what the old hutch python is
+    doing, so I will continue to searching for -I here.
+    """
     domain_re = re.compile('.pcdsn$')
-    ip_re = re.compile(r'^(?:[\d\.]{7,15}|[\w-]+)\s+ami_proxy'
-                       r'\s+.*?\s+-I\s+(?P<ip>\d+\.\d+\.\d+\.\d+)\s+')
+    proxy_re = re.compile(r'ami_proxy.+-I\s+(?P<proxy>\S+)\s')
+    ip_re = re.compile(r'\d+\.\d+\.\d+\.\d+')
     hutch = hutch.lower()
     cnf = CNF.format(hutch)
     procmgr = TOOLS.format('procmgr', 'procmgr')
-    output = call_script([procmgr, 'status', cnf, 'ami_proxy'],
-                         timeout=2,
-                         ignore_return_code=True)
+    output = cache_script([procmgr, 'status', cnf, 'ami_proxy'],
+                          timeout=timeout,
+                          ignore_return_code=True)
     for line in output.split('\n'):
-        ip_match = ip_re.match(line)
-        if ip_match:
-            host, _, _ = socket.gethostbyaddr(ip_match.group('ip'))
-            return domain_re.sub('', host)
+        proxy_match = proxy_re.search(line)
+        if proxy_match:
+            ami_proxy = proxy_match.group('proxy')
+            ip_match = ip_re.match(ami_proxy)
+            if ip_match:
+                domain_name, _, _ = socket.gethostbyaddr(ami_proxy)
+                ami_proxy = domain_re.sub('', domain_name)
+            return ami_proxy